pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orinlab"
version = "1.1.0"
description = "AI Research Lab for Crypto Markets — Telegram bot, signals, on-chain analysis"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["ai", "crypto", "trading", "solana", "telegram", "signals", "defi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.40.0",
    "python-telegram-bot>=21.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "tweepy>=4.14.0",
]

[project.urls]
Homepage = "https://github.com/nujar00t/Orin.LAB"
Repository = "https://github.com/nujar00t/Orin.LAB"
Issues = "https://github.com/nujar00t/Orin.LAB/issues"

[project.scripts]
orinlab = "orinlab.cli:main"

[tool.setuptools.packages.find]
include = ["bot*", "agents*", "utils*", "orinlab*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.example"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["agents", "bot", "utils"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "bot", "utils"]
